requirements.txt
streamlit
pandas
scikit-learn
matplotlib
seaborn
Pillow
